Added NetworkLogger
Daniel started this earlier and experienced problems when trying the new code
that we/I could not explain at that time. But when I came back from the xmas
party I realized what the problem was and started hacking immediately. But it
turned out that major changes were necessary to the Logging subsystem and so
only now this logger is ready.

The NetworkLogger implements the usual Logger interface. A client can
subscribe and will then get all logging messages (almost) instantanteously.
The logging messages can have an arbitrary length as long as the fit into
a FawkesNetworkMessage (the limit being roughly 4 GB here...).

Tim


git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@593 fc82b16a-671c-0410-8732-9903a8d2dd0d