Merged trunk changes r1695:1943 into the trunk
We had left out this merge intentionally last time due to the early stage of
the libgeometry changes. But since a major discussion is currently going on
it's probably a good idea to just throw this at you right now... Features a
Gtk utility API, several new stuff, and unfortunately creates a horrible mess
in many core classes when mixed with trunk and nao changes and after my
midnight merging effort... But with your help this can evolve into a useful
good maintained library. We shouldn't be shy and cut this together to one
consistent, speedy and useful library right now...
The rest of the changes is marginal.

Warning: this is likely to bomb your current stuff, but I see no soft way of
doing this. We probably just have to get through it now.

Tim


git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@1944 fc82b16a-671c-0410-8732-9903a8d2dd0d