Fix the MainLoopAspect, fixes race condition and a deadlock
The MainLoopAspect had a race condition on initialization, if the thread
initialized structures required for the main loop in its init() method a race
condition could cause that mloop() was called too early, possibly causing a
segfault. Since the thread could just idle it could be finalized while the
main loop was running, causing a deadlock or segfault.

+ MainLoopAspect has been transformed. Now there is no MainLoop class anymore.
. Instead the MainLoopAspect thread is used as main thread and woken up,
. eliminating one indirection and allowing for proper blocked finalization.
. ttmainloop plugin changes will make this obvious in just a moment. It cleans
. up the API and is simpler to understand.
+ Use deferred notification of thread startup to actually enable the main loop
. to avoid the initialization race condition.

Tim


git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@2049 fc82b16a-671c-0410-8732-9903a8d2dd0d