Add dummy in FamThread when HAVE_INOTIFY not defined to keep object size
HAVE_INOTIFY might be unset when using the FamThread (if you forget to add
utils.mk for example). This causes an illegal write, because the object is
allocated for the size without the pointer, but the library itself expects
the larger sized object. This can cause memory corruption. Therefore add a
dummy pointer if the define is not set, keeping the object at the same size.