Set SONAME on shared libraries
SONAME is a field in shared object for automatic linker dependency checks.
This is now set by default on shared libraries. By default it is set according
to the Fawkes version. In rare situations this can be overridden per library.
Consult the mailing list before doing such a change to check whether there are
alternative ways to solve it.