Added sysinstall build type for system-wide install, implements ticket #120
The new build mode changes the paths accordingly (and therefore the BT has to
be enabled '''before''' compiling) and it provides targets to install and
uninstall Fawkes system-wide.
The code looks pretty ugly, but that is the price to be paid to have simple
"user" Makefiles with minimal modifications only.

Additionally this commit adapts the interface.mk file to the IFACEDIR change.