Added MemAutoFree utility class
An often encountered problem/annoyance if that temporarily allocated
memory blocks must be freed at the end of init and/or if an exception
is thrown. The std::auto_ptr provides a solution for objects and data
allocated with the new operator, but it does not provide a solution
for memory that needs to be freed with free(). This is now provided
by the MemAutoFree class.