First working version of BBLogger plugin
The BBLogger plugin reads any number of interfaces and writes any data
modifications to a file. This is done either delayed in a concurrent
thread (buffering enabled) or immediately in the event handler.
Buffering needs more RAM as data is copied, and the mode can be harmful
on systems with slow RAM. The in-event writing can influence overall
timing interaction more directly, though may provide better performance
in some situations. It requires some testing per system what works best.

This is work in progress, comments welcome.