Laser: laser beams are now stored counter-clockwise by default
Up to know the data was implicitly assumed to be stored with the
angle growing clockwise. This is incorrect in the Fawkes coordinate
system and is therefore hereby changed.

The new behavior is to store the angles by default with a
counter-clockwise growing angle. If for a laser configuration the
field clockwise_angle is set to true, a second interface named like
the original with an additional " CW" suffix is created to store
clockwise data. The original counter-clockwise data is still provided
as usual.