LaserHT: re-implement line fitting using least squares
Once the dominant line is found, all data points that contributed to
the line during the Hough transform are used to fit a more precise
line. To keep things simple only least squares is used. To avoid
parameter problems with (close to) vertical lines, the coordinate
system is rotated such that the line will be horizontal and thus
eliminating the problem. Proper minimum and maximum ending points
of the line are calculated for fitted line and written to the BB.