BBLogger: allow immediate flushing after each written chunk
There is now a new configuration flag which can be used to turn on
flushing for a logging scenario. This has to be used with care as
this may have severe influence on the system performance. However,
it allows real-time watching of the log file (with the upcoming
bblog tool), so this might be useful when setting up the logging
for a particular scenario for verification.