Main app: get rid of default plugin load exception message
The first experience when running Fawkes is that you get an error
with the exception that the default plugin could not be loaded.
That has bugged me for long enough now. The PluginLoadException
can now report which plugin caused the exception, and the mainapp
ignores the plugin load exception now if the plugin which caused
this exception is "default". Exceptions other than
PluginLoadException are reported as before.