Corrections for proper installation of utils and aspects libraries
The utils library did not define the headers that need to be installed.
The aspects lib headers where installed to an aspects sub-directory
(because the library is named libfawkesaspects), while it should be
named aspect (as the directory in the source tree is named, as it is
mostly used as "include <aspect/logging.h>" etc.).