Use only most specific LDFLAGS variable
Up to now you could set LDFLAGS and LDFLAGS_name (where name is the
binary, lib, or plugin name) to contain linker flags. These were
passed both to the linker. This can cause for example plugin-specific
interfaces to be linked with the flags (for example against the libs)
which were intended for the plugin. This causes unecessary
dependencies. Now LDFLAGS_name is preferred if it exists and used
exclusively. It is checked if it is defined, so it is preferred even
if it simply contains the empty string. If it is not defined, the
LDFLAGS variable is used.