bblogreplay: support blocked timing replay integrated into main loop
There will be scenarios where you want to have the same timing guarantees
for log replay that you have for, e.g., real sensors. This can now be done
by defining a hook at which the log data should be written. For now, one
data item per time will be written, the loop maybe slowed down to replay
time. A future enhancement can be to check at a certain loop if new data
is available and only write if it is, not slowing down that loop. But there
must be something for tomorrow...