bblogreplay: non-blocking blocked timing replay, grace period
In blocked mode data is written at a particular main loop hook. The thread
will block until new data is available.
In non-blocking mode it is only checked if enough time has passed to write
new data. If not the loop just returns without an effect, checking again
in the next iteration.
A grace period has been added. It is used to allow for a small additional
time that the offset and the actual passed time may differ to still be
considered as "close enough" to replay the next data chunk. This is useful
to avoid small offsets to cause large gaps in replay.
Besides that a bug was fixed that cause the loop to wait with the offset
of data element N-1 to wait for the availability of element N.