bblogger: allow turning loggig on/off via BB interface
One of the log threads now becomes the master thread (always the
first created by the plugin). This thread creates a switch interface
which allows turning logging on and off from the outside and
communicates this to the non-master threads.
This is based on a feature request by Christof Rath.