Compatibility with tolua++ 1.0.93
There is a bug in tolua++ 1.0.93 which makes the generator fail when passing
the package file on stdin. We now generate a temporary file that is then
fed to the generator. That is a trivial change.
Basic tests suggest that Fawkes seems to work just fine with both versions.

The CLEAN_FILES make variable can now be set to file (patterns) which should
be removed  on "make clean" in addition to the files which are cleaned anyway.
This is used to remove files generated by tolua++.