Compile Firestation with mirror calibration support if available.
Since the mirror calibration library is not located in the FVBASEDIR
but outside of the Fawkes submodule, the TOP_FVBASEDIR variable was
added that points to the firevision directory in the possibly existing
top-repository. This is then used to determine whether mirror calibration
support is available (HAVE_BULB_CREATOR).

db