refboxcomm: add remote blackboard processor
Added a new refbox processor to access information from a remote
blackboard. The processor will try to maintain an active blackboard
connection to a remote Fawkes and copy the gamestate interface data
from there.
This is mostly intended as a fallback solution when direct
multicasting from the MSL refbox does not work to avoid the problems
we had last time with insufficient network equipment.