fvutils: Colormap::to_image() produces always 512x512 image now
The documentation of Colormap::to_image() stated that the outcome was
always an image of 512x512 pixels. That was not the case, instead it was
always width*2 x height*2, width and height being the width/height
reported by the colormap. In case of the YuvColormap this depends on the
given resolutions and often is smaller. Further down the (code) path this
could lead to problems, presumably also the problems reported in
ticket #131 (to be tested though).

Now we create an image of 512x512 and fill it up, no matter of the actual
size of the colormap, we just assume that the colormap is divided into
proper sections mapping to a full range of 256 entries. This could be
modified to use deepness() (which additionally is a horrible name) later
to get some flexibility back. For now this will do.