firestation: hack fixes crash when mask is loaded before camera images.
The error happens when in load_mask()
1. PNMReader malloc()s its buffer PNM-buffer.
2. YUV-mask-buffer is allocates.
3. PNM-buffer is free()ed.

Even when the complete PNMReader-stuff is erased and instead just
a "void* p = malloc(3*1000*1000)" is inserted and a "free(p)" is appended
to the end of load_mask(), the program crashes when it applies the first
sharpen filter.

However when the free(p) is done before the allocation of the YUV-buffer,
everything is fine! (But we cannot transfer this "solution" to the real
scenario where the PNMReader needs the memory instead of this useless
buffer p).

The hack now changes the order:
1. YUV-mask-buffer is allocates.
2. PNMReader malloc()s its buffer PNM-buffer.
3. PNM-buffer is free()ed.
This works. Unfortunately we don't know the size of the YUV-buffer in
advance in general, so the hack guesses it.

Even if we could know the the YUV-mask-buffer's size before creating
the PNMReader, it would be interesting to know why the free() in the
PNMReader's dtor leads to errors in IPP's sharpen filter.