FuseClient: fix race condition
When a message was enqueued with enqueue_and_wait() immediately after
instantiating the FuseClient (and thus before receiving the greeting
message) the caller was already woken up after the greeting message,
although most likely the expected message has not arrived yet. Wake
the caller now only if a message was received that was given to the
handler for processing.
The problem caused tries to open a camera to fail more often than
justified.
Along the way fixed typo in cams/net.cpp.