firestation: mirrorcalib: fixed lines rotated in opposite direction than slider
When the user pressed 'memorize' (or opened a new file or so), the line
and slider are set to the next expected calibration angle, i.e. to the
current angle + 120 degree or current angle + 120 degree or.
Unfortunately, the slider was set to +120 degrees and the line was
drawn at -120 degrees (from the robot's point of view [in the image, it
was *was* +120 degrees, but the image is mirrored]).