fvcams/v4l2: optionally support using libv4l, fixes for V4L2Camera
Video4Linux2 removes a lot of the conversion code from the kernel and
requires userspace apps to do the conversion (this was different in
v4l1). Thus for many cameras which do not natively stream YUV or RGB
data (and this are most of the cameras) the current v4l2 camera is
useless.
I have added code to V4L2Camera which can optionally use libv4l, a
small library which is meant to solve the very problem of providing
conversion for all kinds of cameras. If the library is detected, it
is used. Otherwise the very same code as before is running. The code
is based on the initial gstreamer patch adding support for libv4l
(released under LGPLv2+).
libv4l is used by VLC, gstreamer and others, so I assume it to be
reasonably stable.

To use the new code you must set format=RGB3 atm. The conversion
unfortunately is slow (~150ms per frame) for the camera I tried
(providing JPEG streams natively), but at least you can get images
again. The other output format which can be emulated by libv4l is
YUV 4:2:0, for which we do not have conversion filters, yet.

Additionally fixed munmap after close bug, and set MMAP as the default
read method. There is no point in defaulting to a method not (yet)
supported.

Thanks to Henk for reporting this problem.