interface generator: change types to stdint.h types
The current types used in interfaces are ambiguous and depend on the
compiler. This is bad when transferring these structs over the network.
Ever tried to open an interface that contains a long int remotely on
a 32 bit system from a 64 bit system or vice versa? Bad things happen.
With the new types the size of all types is well-defined. This also
drops support for number suffixes.