interfaces: added SoccerPenaltyInterface
Moved SplPenaltyInterface from nao repo to trunk repo as
SoccerPenaltyInterface, anticipating that it might be useful for
other leagues as well in the future.

The move is necessary to have all interfaces required by software
in the trunk repo, i.e. the refboxcomm plugin needs the interface
if compiled for operating a robot in the SPL.