SQLite config: fallback to DBs in user homedir
The configuration can now reside in a user's homedir subfolder (defined
as USERDIR). This is required to be able to run a system-wide installed
Fawkes as normal user without requiring to have a world-writable
sysconfdir. The procedure works like this:
- try to open host-specific config from sysconfdir, if that fails, open
  in userdir, if this fails as well throw exception
- search for default.sql in sysconfdir, then in userdir
- attach default.db from sysconfdir, then userdir

This means that the .db files might be created in the userhomedir, while
default.sql is stored in sysconfdir. For now we always try to write to
the default.sql file that has been used to avoid confusion (i.e. if
written back to userdir this would not be used the next time). Still
need to have a graceful error if dumping fails without shouting out
loud.