BuildSys/Interfaces: build generator as presubdir
We had a special case for the interface generator that included a
Makefile snippet either in the generator dir or its parent interfaces
dir. This was done in the past for dependency reasons, but it causes an
awkward case when it comes to man pages. Therefore made the generator
dir a proper self-contained module and make the interfaces depend on the
subdir build. The dir is build as presubdir to properly build before the
interfaces.