AspectProviderAspect: support adding new aspects at run-time
Up to now aspects were static in the sense that they were added once
during development and could then be used. But aspects could not be
added at run-time, for example by a plugin providing a new aspect.

This patch adds the AspectProviderAspect, which allows a thread to
provide a new aspect which in turn other threads can acquire.
The system asserts that an aspect provider is not removed while there
still is a thread in the system that has the provided aspect.

The aspect is a new standard aspect initialized automatically on
startup. Test plugins have been written and I can share them. Use this
feature rarely! Use regular means of communication via the blackboard
and only provide an aspect if you are certain it cannot be accomplished
in a better or similarly good way than with a new aspect.

This feature is added for two reasons: first it has bugged me for two
years, and I had a start of the solution in my repo for more than a
year, and second I might encounter a situation were I need this soon. So
it is not a feature added for wide use.