plugin lib: unload plugins in reverse order, now really
The intention of the existing code was to unload all remaining plugins
in the reverse order they were loaded during destruction, i.e. during
Fawkes shutdown. This was meant to maximize the changes of properly
finalizing threads, because they could have relied on features provided
by plugins loaded earlier.

But a map was used to store the plugins, which is a *sorted* associative
container, thus the plugins were unloaded in reverse alphabetical order.
In my extensible aspect test case this problem became obvious. We now
use a list to store the plugins which is unsorted by default. This way
we get the ordering that was intended.