webview: use new WebviewAspect
Adapted the webview plugin to recent libfawkeswebview changes and use
the WebviewAspect. This way plugins can now arbitrarily add web request
processors which will be executed by webview. Since the URL manager is
maintained by the aspect initializer/finalizer webview does not have to
be loaded if a plugin with a thread with the WebviewAspect is loaded,
but the processor is still called if webview is loaded later.