Preliminary support for the Kinect camera.
The implementation is based on the freenect driver. Since neither the
driver nor libusb-1.0 which is required by the driver are currently
available as a RPM in Fedora you have to install those from source and
patch the Makefiles.

Currently, the Firevision Kinect camera allows to receive the color
image as well as a false color viusalization of the depth image. In
order to obtain real depth values from the camera some kind of
calibration is necessary.

What I noticed is that the driver sucks up quite some CPU
cycles---around 60% CPU usage when images are fetched from the camera.
Doesn't seem necessary but I haven't looked at it, yet.