mongodb: MongoDBAspect allow to name a MongoDB client config
Threads with the MongoDB aspect can now specify a MongoDB configuration
name which is used to setup the MongoDB connection. The authentication
data must then be present in the configuration, rather than being passed
to the aspect ctor. This makes it possible to vary settings and to
support completely different database settings and connections for
different threads.