buildsys: enforce linkage for undefined symbols
Until now it was ok to use symbols of libraries that we did not
explicitly link against in plugins and shared libraries. This caused
plugins to not load sometimes because symbols were missing. Now we
enforce that all shared libraries are indeed linked against every
library from which symbols are used.

This might also fix the problem that plugins were sometimes not properly
unloaded.

Fixed all active plugins and shared libraries in the main tree.