webview: introduce navigation manager
Introduced a navigation manager, which can be used to add navigation
entries to the webview navbar. This was buried in the webview plugin
before. Similarly to the URL manager, the nav manager exists static and
can be used even in absence of the webview plugin, further keeping the
WebviewAspect as an optional aspect.