Support building without webview/libmicrohttpd
The software should work without libmicrohttpd, therefore made changes
to accomplish that. The WebviewAspect is now built conditionally, and
plugins should be omitted or make the use of the WebviewAspect optional
(depending on the degree on which they rely on the aspect to work).
Along the way added a missing include pulled in by newer versions of
rrdtool but not older ones.