rrd: improve description memory handling
RRD descriptions were keeping pointers and references to most stuff.
This could cause problem if objects got erased. We now copy a lot more.
The RRDManager can now configure RRDDefinitions to get notified on
destruction as to remove the RRD if it is deleted. Keep graph elements
as pointers, required because of virtual inheritance.