ThreadList: proper finalization on failed initialization
Threads in a ThreadList were not all properly finalized. Although fully
initialized threads were properly finalized on an error of a thread, the
thread causing the abort was not finalized. Although this was intended
for the thread's finalize method, already fully initialized aspects were
not finalized.

Consider a thread with a LoggerAspect. The logger is already registered
by the aspect and therefore used. If init() now throws an Exception, the
aspect was not finalized, and hence the no longer existing logger was
not removed, causing an error later, usually a "pure virtual method
called" error.

Now the finalizer is called, if it was not the initialized which failed
(in which case we need to rely on the initializer to finalize the
already existed aspects).