roomba: native types and new messages in interface
Instead of forcing SI units stick to the Roomba units for the
Roomba500Interface. It is meant for direct instruction anyway. Later,
when providing a motor interface or the like this will be in meters.

Added more actuator command messages like setting motors, driving,
and docking. Also support setting the mode over the interface.

Adapt Roomba500 and thread classes to these changes.