joystick/JoystickInterface: separate axes
A relict from RCSoft was that the joystick data was read, and then X/Y
axes were assumed for two consecutive values. The Linux joystick API
does not make that distinction, and the Logitech Rumblepad, for example,
has two axes for an X/Y combination, then one axis for a slider, and
then two X/Y coupled values again, rendering the scheme absurd.

The new way is to have a single axis array to which we write through all
axis values, not imposing assumptions onto the values.

Adapt default joystick path to changed kernel default, it's
/dev/input/js0 now. Fix crasher (uninitialized variable).