bbsync: change remapping definition
Remapping used to be defined as "local ID = remote ID", no matter if
local was reading or writing. This does not seem particularly intuitive.
Changed so it reads "reading ID = writing ID" now. This mapping seems to
be easier to understand. It is documented this way at wiki:Plugins/bbsync.