RemoteBlackBoard: guard against some typical deadlocks
Some operations, like opening interfaces in an event handler cannot be
done, because they cause a deadlock (network receiver thread calls event
handler, calls open interface, sends message and waits for reply, which
cannot be received because we're doing this from within the receiving
routine). A simple check should catch this most of the times.