laser-filter: LaserDataFilter API multi-array support
Changed the laser data filter API to be able to process multiple input
arrays at the same time. This is required for the upcoming merge filter.
Also change some assumptions about the persistence of input and output
interfaces to allow for simpler and more flexible chaining of filters.
Adapt all existing filters to the changes.