laser: auto-detect non-chardev mode PCAN driver
If the PCAN driver has been compiled in netdev instead of chardev mode
the first read will simply stall. This bug is well documented in the
PCAN documentation...
By reading the driver info from the /proc file system we can detect the
netdev mode and throw an error.

Besides that fix typo in an error message and shorten the plugin
description.