lasergui: support selecting of interface
Instead of always showing the Laser360Interface or Laser720Interface
with the ID "Laser" provide a way to choose the interface from the list
of interfaces using the new InterfaceChooserDialog. The highres/lowres
button has been removed, the way to switch is now to open either the 360
or 720 value interface. Choosing the interface is particularly useful on
systems with multiple laser range finders and with the laser-filter
plugin and intermediate result interfaces.