laser: support URG identification via USB bus/device number
With multiple Hokuyos on the robot, and the URG not providing a serial
number the proper USB way, we need to have different means to identify a
laser range finder. To avoid randomly opening laser range finders the
USB bus number and device number can now be configured which is then
used to identify the Hokuyo. You can determine the Hokuyo's bus and
device number like this:
lsusb -t
Look for "Vendor 0x15d1 Product 0x0000". Go up one entry to get the
device name, and one more to get the bus name, e.g. if the output is:
Bus#  4
`-Dev#   1 Vendor 0x1d6b Product 0x0001
  `-Dev#   6 Vendor 0x15d1 Product 0x0000

The busnum is 4, the device num is 1. This device corresponds to the USB
port and is stable, while the Hokuyo's device num changes on each
replugging.