laser: support probing for URG by serial number
Probing for URG by bus and device number turned out to be useless if
devices are connected to the same hub and hence bus.

This version can probe for the serial number. Unfortunately, the Hokuyos
do not provide their serial number via the appropriate USB field, which
means that we need to open each Hokuyo and query the information. We use
explicit locking on the device file to ensure that there are not two
acquisition threads interfering with each other by communicating with
the same device.

The thoughtful reviewer might suspect a race condition in the
initialization, if the actual user of a device cannot lock it while
another thread has it locked for probing. This is not the case because
Fawkes initializes the threads one after another and not in parallel.