openni-pointcloud: calc and publish Kinect point clouds
This plugin reads the depth image and converts it to cartesian
coordinates which are published using a SharedMemoryImageBuffer for now,
until we have something useful in the blackboard (depends on ticket #61).

The code is currently untested because I don't have a Kinect here. Needs
a little demo app or a plugin that uses the data.