pantilt: fix echo cancellation
RS485 adapters seem to have the property of an inherent echo, which can
only be turned off for some models, but not ours. That was anticipated
by simply reading the echo off the line after sending. Obviously, it
seems that it once was enough to just execute a simple read. Not anymore,
we need to loop until we got all the nasty echo bytes. It's the only way
to get rid of the protocol errors we were seeing recently.