openni point cloud viewer: simple OpenGL app to visualize PCLs
To debug a nasty problem in the cannikin cup detection plugin added a
simple visualization that shows the point clouds gathered via the
openni-pointcloud plugin. It also shows a half-circle in 1m distance to
get a better glance at the data, a red dot for the camera position, and
a green spot for a detected object, if connected to a remote blackboard
and a writer exists for the ObjectPositionInterface (ID can be passed on
the command line).