openni: run and maintain XnSensorServer
The PrimeSense sensor module starts a separate process to interact
with the camera, but the process is not closed. Rather it waits for a 10
second timeout of no client connections and then closes. Since the fork
and execution is poorly implemented, remnants like the Fawkes network
port stay active in the XnSensorServer and cause consecutive restarts of
Fawkes to fail. The workaround is to manually "killall XnSensorServer"
after stopping Fawkes or to wait 10 seconds for the timeout.

This little patch enables the context thread to run the sensor server by
itself, allowing for a proper and immediate shutdown on unloading the
openni plugin (or closing Fawkes).