laser-filters: projection: LEFT parameter is now ROT_LASER
Laser projection is done in three steps:
(1) the laser coordinate system is rotated so that it is the
coordinate system of the laser body.
(2) the laser body is rotated along the fixtures so that the
X,Y,Z axis are parallel to the Fawkes coordinate system.
(3) the rotated laser body is translated into the Fawkes
coordinate system.

Up to now, the behavior of the first step depends on the config value
LEFT.
If LEFT=true, Z was expected to be rotated by 90 deg and Y by -90 deg.
If LEFT=false, Z was expected to be rotated by -90 deg and Y by -90 deg.

Now, the first step is completely configurable. The configuration values
x_rot_laser, y_rot_laser, z_rot_laser are expected to be the rotation of
the laser inside the laser body.

Note: For better readability, the configuration values x_rot, y_rot,
z_rot are renamed to x_rot_fixture, y_rot_fixture, z_rot_fixture.