openni: fix compile error if OpenNI is not available
Split libfawkesopenniutils in an OpenNI dependend part and a client
utils library which only depends on Fawkes. This way, the visualizations
can be built even though OpenNI is not present, which is useful because
OpenNI might only be necessary on the robot with the Kinect, but the
visualization is run remotely.