libfawkesbaseapp: move code from mainapp to new lib
Basically all code from mainapp has been moved to the new baseapp
library to make it easy to embed Fawkes into other systems, programs, and
frameworks.

Some of the code from the main thread has been moved to a new
fawkes::runtime namespace, where it is easier accessible and it moves
some of the parameter out of the main thread where it did not really
belong.

Implements ticket #140.