Interface: add support for side store buffers
Interfaces now support additionaly copy buffers which can be used to
store intermediate versions of the interface data.One example use case
is to save the current shared memory content at one point in time at a
specific main loop hook, and restore it only later at a suitable time in
another continuous thread. Another useful application is to keep a
history for hysteresis processing, or to observe the development of the
values in an interface.