Lua utils: interface importer supports two-stage mode
The interface importer can now handle reading interfaces in a two-stage
mode, where it provides method to read the current blackboard content to
buffers, and restores it to the private sections of the interfaces only
on specific request.

This is useful to update a set of interfaces in a specific hook, but
update to the new content only at a specific point in time, getting the
then-latest content.