luaagent: support continuous execution mode
With this change we can now support agents which assume to run their own
main loop inside the Lua code. C functions for triggering interface
updates are provided and reload support still works by wrapping the
execution in a thread which can be cancelled.

This mode is not encouraged, but sometimes necessary.