netcomm: fix possible hangs/deadlocks in AvahiThread
A problem could arise that communication with Avahi would hang. I
tracked this down and removing pending services properly seems to fix
the problem. This can introduce a race condition with a deadlock, if the
thread was cancelled while pending services were removed. The mutex
protecting the pending service queue would not be unlocked in that case.
We prevent this by disallowing cancelling during that time which is
rather short.