libfawkeslogging: introducing new lib to concentrate logging
The code for logging was spread between utils and netcomm. Putting this
into its own library will beautify things a bit. This also allows us to
move the LoggerEmployer out of the aspects library and make multi logger
an employer, which will allow us to trim down the main thread.