Dynamic modules: reorganized code and argument passing
The ModuleManager and its surroundings were overly general and complex.
Past experience has shown that we don't need the complexity to adapt to
different platforms. The final kick for the decision was the need to
pass different flags for loading modules in some situations.
Adapted the code that used the plugin manager (factory).