naoqi: Fawkes plugin and NaoQi module to interact with NaoQi
The NaoQiAspect provides access to a shared NaoQi broker which allows to
interact with NaoQi by using its services or reading from ALMemory. To
gain access to the broker, the naoqi plugin provides this aspect to other
threads or plugins. Currently, the plugin expects to be embedded into
NaoQi, later we may add embedding of NaoQi in the plugin. The NaoQi
module naofawkes is loaded in NaoQi and it embeds a Fawkes
instance, sets up the shared broker and loads the nao plugin.

The NaoHardwareInterface has been added from the old fawkes-nao domain
repository and code is upcoming to fill it from within the embedded
Fawkes instance.