Nao interfaces: split NaoHardwareInterface into smaller chunks
The NaoHardwareInterface has grown over the years, making it less
efficient and confusing. There are now three interfaces replacing it,
first the NaoJointPositionInterface (taking over the original intent of
the NaoHardwareInterface), the NaoJointStiffnessInterface (for which we
used a second instance of the NaoHardwareInterface as a workaround), and
the NaoSensorInterface transporting all the sensor data of the Nao.