naoqi: DCM thread handles servo movements, interface split
The DCM thread can now issue servo movements. Due to deficiencies in
ALMotion (it does not update its internal belief about the robot's
posture and stiffness) we use ALMotion for setting stiffness. Adapt to
the split of NaoHardwareInterface. We also provide two version of each
interface, one that is filled during the sensor hook, and one that is
filled at full DCM rate, but only if there is a reader for the
interface.