naoqi-motion: plugin to call ALMotion
Initial version of naoqi-motion plugin to instruct ALMotion. It is still
incomplete, but can already make the robot standup or walk.
Note, that due to a problem in ALMotion setting servos via DCM and then
calling ALMotion for these servos might make the robot snap back to the
position it had before the DCM motions, because the internal model of
ALMotion is not updated. The fix_angles() method should help, but it
does not work as expected, workaround has been requested in Aldebaran's
forum.