Interface: compare buffers to private memory
Added Interface::compare_buffers() that allows to compare a private
buffer to the private memory area. This is useful for a new use case of
private buffers: stage away the recent memory content, set new
(computed) values, and then compare buffers if anything has actually
changed. Then, only write the data if it has. This saves bandwidth and
power.