NaoJointPositionInterface: removed Goto, added Move messages
The joint position interface can be used for quick timing joint
movements, and to enqueue trajectories timed SetServo commands are
useful. However, the Goto messages were more akin to motion messages
(and in fact executed via ALMotion). But we have seen that for smooth
movements commands parameterized by speed rather than time are more
useful. We follow the recently modified HumanoidMotionInterface in these
semantics and add MoveServo(s) messages, which move a servo at a given
speed. It is the duty of the caller to stop it appropriately. This also
has the advantage that this allows for better blending of replacing
messages. Remove the goto family of messages.