nao: adapt DCM thread to interface changes
Especially handle the Move-style messages and remove the old Goto
messages. This supports smooth blending of different move commands. It
is not supported to mix move and set commands at the same time atm.
Also, ALMotion does not properly update its internal state when writing
to DCM directly, so one should either use the one or the other kind of
commands. As soon as naoqi-motion is used, only the Move style messages
should be used for servo movements to avoid problems until ALMotion is
fixed.