libfawkesbaseapp: add additional default plugin init option
An additional default plugin can be given, that is loaded regardless of
command line parameters or init options. It is intended to serve as a
platform-specific default plugin, e.g. one that is required to make the
platform usable, especially when integrating with another software
system like on the Nao.