libfawkeslogging: add SyslogLogger
New logger to log to syslog. Unfortunately syslog is not anywhere close
to thread-safe, so we modify the ident string of the whole system, for
example causing NaoQi log messages to appear with "Fawkes" ident prefix.
But we at least can now log to the same location and avoid log
scattering for basic messages.