libfawkesbaseapp: show forced exit msg only after 3 sec
When Ctrl-C is pressed we immediately printed a message that described
how to achieve a forced exit by pressing Ctrl-C again. Most of the time
this is not necessary though, and we print the message and immediately
exit. To make this nicer setup an SIGALRM and print the message only if
the finalization/shutdown procedure takes three seconds or more.