buildsys: make linker configurable
Currently we used $(CC) to define the linker. For some situations, for
example when linking CUDA code into a plugin, it is required to specify
a different linker. To be able to do this independently of the compiler
and to do this for only a specific plugin, library, or binary make the
linker a setting which can be easily overridden.

It is now possible to specify LD_[target] where target is the same that
you define as OBJS_[target].