Move interfaces to libs dir
We have had trouble for a long time for that no library in libs could
use interfaces because they were built only later.

Today, we have finally moved interfaces to src/libs/interfaces. Even
though it gives a slightly lesser highlight on interfaces it is the
pragmatic approach to building more sophisticated libraries. It should
still be the goal to have libraries independent of the communication,
but for some things is really useful or outright required.