fawkestf: add concept of disabling listener/publisher
The listener and publisher can now be disabled, i.e. instantiated
without a BB. This is useful for the TransformAspect to allow for having
only either listener or publisher, but not necessarily both. With a
disabled listener/publisher you get an exception with a useful message,
rather than just a segfault. The new concept makes the
TransformPublisherProtector superfluous which served the same purpose.