static-transforms: allow partial configurations
Translations and rotations are optional and default to zero translation
or rotation if not given. If any one value is given for a translation or
rotation though it must be completely specified or it is an error.
Rotations can still be configured by either specifying the quaternion or
the Euler angles.